fix bug unique key
